% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banknote.R
\docType{data}
\name{banknote}
\alias{banknote}
\title{Swiss banknotes data}
\format{
A data frame with 200 rows and 7 columns:
\describe{
\item{Status}{the status of the banknote: \code{genuine} or \code{counterfeit}}
\item{Length}{Length of bill (mm)}
\item{Left}{Width of left edge (mm)}
\item{Right}{Width of right edge (mm)}
\item{Bottom}{Bottom margin width (mm)}
\item{Top}{Top margin width (mm)}
\item{Diagonal}{Length of diagonal (mm)}
}
}
\source{
Flury, B. and Riedwyl, H. (1988). Multivariate Statistics: A practical approach. London: Chapman & Hall, Tables 1.1 and 1.2, pp. 5-8.
}
\usage{
banknote
}
\description{
The data set contains six measurements made on 100 genuine and 100 counterfeit old-Swiss 1000-franc bank notes.
The data frame and the documentation is a copy of \link[mclust:banknote]{mclust::banknote}.
}
\keyword{datasets}
