% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_n_primes.R
\name{generate_n_primes}
\alias{generate_n_primes}
\title{Generate a Sequence of Prime Numbers}
\usage{
generate_n_primes(n, one = FALSE)
}
\arguments{
\item{n}{the number of primes to generate.}

\item{one}{should \code{1} included in the sequence (default: \code{FALSE})}
}
\value{
an integer vector of prime numbers
}
\description{
Generates a vector of the first \code{n} primes using \code{\link[gmp:nextprime]{gmp::nextprime()}}.
}
\examples{
generate_n_primes(5)
generate_n_primes(5, TRUE)
}
