% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(ar)
}
\arguments{
\item{ar}{numeric variable.}
}
\value{
Returns normalized variable.
}
\description{
Normalization of variable.
}
\details{
Normalization of variable: \code{ar<-(ar-min(ar))/(max(ar)-min(ar))}
}
\examples{
normalize(iris[,1])
}
\author{
Jaroslav Myslivec \href{mailto:jaroslav.myslivec@upce.cz}{jaroslav.myslivec@upce.cz}, Sigbert Klinke \href{mailto:sigbert@hu-berlin.de}{sigbert@hu-berlin.de}
}
