\name{numarray}
\Rdversion{1.1}
\alias{numarray}
\title{Numeric array}
\description{
Extracts numeric array from data frame.
}
\usage{
numarray(df)
}
\arguments{
  \item{df}{data frame.}
}
\details{
Extracts numeric array from data frame.
}
\value{
Returns numeric array.
}
\author{Jaroslav Myslivec <jaroslav.myslivec@upce.cz>}
