% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{zzz}
\alias{zzz}
\title{Comparison}
\usage{
zzz()
}
\value{
nothing
}
\description{
Creates and displays a temporary PDF file with different diagrams comparing \code{andrews} and \code{andrews0} plots.
}
\examples{
if (interactive()) zzz()
}
