% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/numarray.R
\name{numarray}
\alias{numarray}
\title{Numeric array}
\usage{
numarray(df)
}
\arguments{
\item{df}{data frame.}
}
\value{
Returns numeric array.
}
\description{
Extracts numeric array from data frame.
}
\details{
Extracts numeric array from data frame.
}
\examples{
numarray(iris)
}
\author{
Jaroslav Myslivec \href{mailto:jaroslav.myslivec@upce.cz}{jaroslav.myslivec@upce.cz}, Sigbert Klinke \href{mailto:sigbert@hu-berlin.de}{sigbert@hu-berlin.de}
}
