% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectand.R
\name{selectand}
\alias{selectand}
\title{Selecting in Andrews curves}
\usage{
selectand(df, type = 1, step = 100, ncol = 0, from = 0, to = 1, col = 2)
}
\arguments{
\item{df}{data frame.}

\item{type}{type of curve.}

\item{step}{smoothness of curves.}

\item{ncol}{number of column in data frame for selection.}

\item{from}{from value.}

\item{to}{to value.}

\item{col}{color of selected objects.}
}
\value{
Nothing
}
\description{
Selecting object utility in Andrews curves
}
\details{
Define which objects will be selected (colored) in Andrews curves.
}
\examples{
data(iris)
andrews(iris,clr=5,ymax=3)
selectand(iris,ncol=1,from=5,to=5.5,col=1)
}
\author{
Jaroslav Myslivec \href{mailto:jaroslav.myslivec@upce.cz}{jaroslav.myslivec@upce.cz}
}
\keyword{hplot}
